Move display code into display.rs
