+++ rust/datafusion/src/logical_plan/mod.rs
@@ -36,10 +36,13 @@ use crate::{
 use crate::{physical_plan::udf::ScalarUDF, sql::parser::FileType};
 use arrow::record_batch::RecordBatch;
 
+mod display;
 mod expr;
 mod extension;
 mod operators;
 
+pub use display::display_schema;
+use display::{GraphvizVisitor, IndentVisitor};
 pub use expr::{
     and, array, avg, binary_expr, col, concat, count, create_udaf, create_udf,
     exprlist_to_fields, length, lit, max, min, sum, Expr, Literal,
@@ -335,215 +338,6 @@ impl LogicalPlan {
     }
 }
 
-/// Formats plans with a single line per node. For example:
-///
-/// Projection: #id
-///    Filter: #state Eq Utf8(\"CO\")\
-///       CsvScan: employee.csv projection=Some([0, 3])";
-struct IndentVisitor<'a, 'b> {
-    f: &'a mut fmt::Formatter<'b>,
-    /// If true, includes summarized schema information
-    with_schema: bool,
-    indent: u32,
-}
-
-impl<'a, 'b> IndentVisitor<'a, 'b> {
-    fn write_indent(&mut self) -> fmt::Result {
-        for _ in 0..self.indent {
-            write!(self.f, "  ")?;
-        }
-        Ok(())
-    }
-}
-
-impl<'a, 'b> PlanVisitor for IndentVisitor<'a, 'b> {
-    type Error = fmt::Error;
-
-    fn pre_visit(&mut self, plan: &LogicalPlan) -> std::result::Result<bool, fmt::Error> {
-        if self.indent > 0 {
-            writeln!(self.f)?;
-        }
-        self.write_indent()?;
-
-        write!(self.f, "{}", plan.display())?;
-        if self.with_schema {
-            write!(self.f, " {}", display_schema(plan.schema()))?;
-        }
-
-        self.indent += 1;
-        Ok(true)
-    }
-
-    fn post_visit(
-        &mut self,
-        _plan: &LogicalPlan,
-    ) -> std::result::Result<bool, fmt::Error> {
-        self.indent -= 1;
-        Ok(true)
-    }
-}
-
-/// Print the schema in a compact representation to `buf`
-///
-/// For example: `foo:Utf8` if `foo` can not be null, and
-/// `foo:Utf8;N` if `foo` is nullable.
-///
-/// ```
-/// use arrow::datatypes::{Field, Schema, DataType};
-/// # use datafusion::logical_plan::display_schema;
-/// let schema = Schema::new(vec![
-///     Field::new("id", DataType::Int32, false),
-///     Field::new("first_name", DataType::Utf8, true),
-///  ]);
-///
-///  assert_eq!(
-///      "[id:Int32, first_name:Utf8;N]",
-///      format!("{}", display_schema(&schema))
-///  );
-/// ```
-pub fn display_schema<'a>(schema: &'a Schema) -> impl fmt::Display + 'a {
-    struct Wrapper<'a>(&'a Schema);
-
-    impl<'a> fmt::Display for Wrapper<'a> {
-        fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-            write!(f, "[")?;
-            for (idx, field) in self.0.fields().iter().enumerate() {
-                if idx > 0 {
-                    write!(f, ", ")?;
-                }
-                let nullable_str = if field.is_nullable() { ";N" } else { "" };
-                write!(
-                    f,
-                    "{}:{:?}{}",
-                    field.name(),
-                    field.data_type(),
-                    nullable_str
-                )?;
-            }
-            write!(f, "]")
-        }
-    }
-    Wrapper(schema)
-}
-
-/// Logic related to creating DOT language graphs.
-#[derive(Default)]
-struct GraphvizBuilder {
-    id_gen: usize,
-}
-
-impl GraphvizBuilder {
-    fn next_id(&mut self) -> usize {
-        self.id_gen += 1;
-        self.id_gen
-    }
-
-    // write out the start of the subgraph cluster
-    fn start_cluster(&mut self, f: &mut fmt::Formatter, title: &str) -> fmt::Result {
-        writeln!(f, "  subgraph cluster_{}", self.next_id())?;
-        writeln!(f, "  {{")?;
-        writeln!(f, "    graph[label={}]", Self::quoted(title))
-    }
-
-    // write out the end of the subgraph cluster
-    fn end_cluster(&mut self, f: &mut fmt::Formatter) -> fmt::Result {
-        writeln!(f, "  }}")
-    }
-
-    /// makes a quoted string suitable for inclusion in a graphviz chart
-    fn quoted(label: &str) -> String {
-        let label = label.replace('"', "_");
-        format!("\"{}\"", label)
-    }
-}
-
-/// Formats plans for graphical display using the `DOT` language. This
-/// format can be visualized using software from
-/// [`graphviz`](https://graphviz.org/)
-struct GraphvizVisitor<'a, 'b> {
-    f: &'a mut fmt::Formatter<'b>,
-    graphviz_builder: GraphvizBuilder,
-    /// If true, includes summarized schema information
-    with_schema: bool,
-
-    /// Holds the ids (as generated from `graphviz_builder` of all
-    /// parent nodes
-    parent_ids: Vec<usize>,
-}
-
-impl<'a, 'b> GraphvizVisitor<'a, 'b> {
-    fn new(f: &'a mut fmt::Formatter<'b>) -> Self {
-        Self {
-            f,
-            graphviz_builder: GraphvizBuilder::default(),
-            with_schema: false,
-            parent_ids: Vec::new(),
-        }
-    }
-
-    /// Sets a flag which controls if the output schema is displayed
-    fn set_with_schema(&mut self, with_schema: bool) {
-        self.with_schema = with_schema;
-    }
-
-    fn pre_visit_plan(&mut self, label: &str) -> fmt::Result {
-        self.graphviz_builder.start_cluster(self.f, label)
-    }
-
-    fn post_visit_plan(&mut self) -> fmt::Result {
-        self.graphviz_builder.end_cluster(self.f)
-    }
-}
-
-impl<'a, 'b> PlanVisitor for GraphvizVisitor<'a, 'b> {
-    type Error = fmt::Error;
-
-    fn pre_visit(&mut self, plan: &LogicalPlan) -> std::result::Result<bool, fmt::Error> {
-        let id = self.graphviz_builder.next_id();
-
-        // Create a new graph node for `plan` such as
-        // id [label="foo"]
-        let label = if self.with_schema {
-            format!(
-                "{}\\nSchema: {}",
-                plan.display(),
-                display_schema(plan.schema())
-            )
-        } else {
-            format!("{}", plan.display())
-        };
-
-        writeln!(
-            self.f,
-            "    {}[shape=box label={}]",
-            id,
-            GraphvizBuilder::quoted(&label)
-        )?;
-
-        // Create an edge to our parent node, if any
-        //  parent_id -> id
-        if let Some(parent_id) = self.parent_ids.last() {
-            writeln!(
-                self.f,
-                "    {} -> {} [arrowhead=none, arrowtail=normal, dir=back]",
-                parent_id, id
-            )?;
-        }
-
-        self.parent_ids.push(id);
-        Ok(true)
-    }
-
-    fn post_visit(
-        &mut self,
-        _plan: &LogicalPlan,
-    ) -> std::result::Result<bool, fmt::Error> {
-        // always be non-empty as pre_visit always pushes
-        self.parent_ids.pop().unwrap();
-        Ok(true)
-    }
-}
-
 // Various implementations for printing out LogicalPlans
 impl LogicalPlan {
     /// Return a `format`able structure that produces a single line
@@ -579,11 +373,7 @@ impl LogicalPlan {
         impl<'a> fmt::Display for Wrapper<'a> {
             fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                 let with_schema = false;
-                let mut visitor = IndentVisitor {
-                    f,
-                    with_schema,
-                    indent: 0,
-                };
+                let mut visitor = IndentVisitor::new(f, with_schema);
                 self.0.accept(&mut visitor).unwrap();
                 Ok(())
             }
@@ -624,11 +414,7 @@ impl LogicalPlan {
         impl<'a> fmt::Display for Wrapper<'a> {
             fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
                 let with_schema = true;
-                let mut visitor = IndentVisitor {
-                    f,
-                    with_schema,
-                    indent: 0,
-                };
+                let mut visitor = IndentVisitor::new(f, with_schema);
                 self.0.accept(&mut visitor).unwrap();
                 Ok(())
             }
@@ -1262,31 +1048,6 @@ mod tests {
         assert!(!stringified_plan.should_display(false));
 
         Ok(())
-    }
-
-    #[test]
-    fn test_visitor() {
-        let schema = Schema::new(vec![]);
-        assert_eq!("[]", format!("{}", display_schema(&schema)));
-    }
-
-    #[test]
-    fn test_display_empty_schema() {
-        let schema = Schema::new(vec![]);
-        assert_eq!("[]", format!("{}", display_schema(&schema)));
-    }
-
-    #[test]
-    fn test_display_schema() {
-        let schema = Schema::new(vec![
-            Field::new("id", DataType::Int32, false),
-            Field::new("first_name", DataType::Utf8, true),
-        ]);
-
-        assert_eq!(
-            "[id:Int32, first_name:Utf8;N]",
-            format!("{}", display_schema(&schema))
-        );
     }
 
     fn display_plan() -> LogicalPlan {
